This PR is related to https://github.com/extism/extism/pull/605

```
PS D:\dylibso\go-sdk> go test -benchmem -run=^$ -bench .
goos: windows
goarch: amd64
pkg: github.com/extism/go-sdk
cpu: 13th Gen Intel(R) Core(TM) i7-1365U
BenchmarkInitialize/noop-12                         385           3051739 ns/op         2942236 B/op       2751 allocs/op
BenchmarkInitializeWithCache/noop-12                2112            504269 ns/op         1686493 B/op       1554 allocs/op
```

Without using compilation cache: 3ms
with compilation cache: 0.5ms

Since we're already exposing `wazero.RuntimeConfig`, I think we just need to mention it in the README