@nilslice 2023-11-29
Great work! I'm curious what the lift would be pre 1.0 to actually change the SDK API to make it so we don't actually expose any wazero functions or package info -- instead we add some configuration functions to our API that set thing in wazero. 

Just as we don't expose anything wasmtime related in the Rust SDK.. if it's a massive change that would also be good to know.