@mhmd-azeez 2023-11-30
@nilslice  Currently we are exposing these wazero APIs:
- [`RuntimeConfig`](https://github.com/tetratelabs/wazero/blob/d39d84505ad8021b5ba47bbc9dd55911b9b0406f/config.go#L40-L171)
   - `WithCoreFeatures`: ❌
   - `WithMemoryLimitPages`: ✅
   - `WithMemoryCapacityFromMax`: ❌
   - `WithDebugInfoEnabled`: ❌
   - `WithCompilationCache`: ❌
   - `WithCustomSections`: ❌
   - `WithCloseOnContextDone`: ❓ (If timeout is configured, this will automatically be enabled)
- [`ModuleConfig`](https://github.com/tetratelabs/wazero/blob/d39d84505ad8021b5ba47bbc9dd55911b9b0406f/config.go#L444-L638)
   - `WithArgs`: ❌
   - `WithEnv`: ❌
   - `WithFS`: ✅
   - `WithFSConfig`: ✅ (wazero allows readonly mounting too)
   - `WithName`: ✅
   - `WithStartFunctions` ✅ (we have our own module initialization logic)
   - `WithStderr`: ❌
   - `WithStdin`: ❌
   - `WithStdout`: ❌
   - `WithWalltime`: ❌
   - `WithSysWalltime`: ❌
   - `WithNanotime`: ❌
   - `WithSysNanotime`: ❌
   - `WithNanosleep`: ❌
   - `WithOsyield`: ❌
   - `WithSysNanosleep`: ❌
   - `WithRandSource`: ❌
- [`api.ValueType`](https://github.com/tetratelabs/wazero/blob/d39d84505ad8021b5ba47bbc9dd55911b9b0406f/api/wasm.go#L83-L110)
- [`api.EncodeXX/api.DecodeXX`](https://github.com/tetratelabs/wazero/blob/d39d84505ad8021b5ba47bbc9dd55911b9b0406f/api/wasm.go#L693-L755)
- [`api.Memory`](https://github.com/tetratelabs/wazero/blob/d39d84505ad8021b5ba47bbc9dd55911b9b0406f/api/wasm.go#L557-L673)

It all comes down to how many of these we want to implement for 1.0, maybe we can start with a minimal number of these implemented and then add them as people ask for them